+++ packages/patternfly-4/react-charts/src/components/ChartTheme/themes/donut-threshold-theme.ts
@@ -1,22 +1,39 @@
-// TODO Replace label props with PF css when available
+/* eslint-disable camelcase */
+import {
+  chart_donut_threshold_first_Color,
+  chart_donut_threshold_second_Color,
+  chart_donut_threshold_third_Color,
+  chart_donut_threshold_dynamic_pie_Height,
+  chart_donut_threshold_dynamic_pie_Width,
+  chart_donut_threshold_static_pie_Height,
+  chart_donut_threshold_static_pie_angle_Padding,
+  chart_donut_threshold_static_pie_Width,
+} from '@patternfly/react-tokens';
 
 // Donut threshold dynamic theme
 export const DonutThresholdDynamicTheme = {
   legend: {
-    colorScale: ['#D2D2D2', '#BBBBBB']
+    colorScale: [
+      chart_donut_threshold_second_Color.value,
+      chart_donut_threshold_third_Color.value
+    ]
   },
   pie: {
-    height: 202,
-    width: 202
+    height: chart_donut_threshold_dynamic_pie_Height.value,
+    width: chart_donut_threshold_dynamic_pie_Width.value
   }
 };
 
 // Donut threshold static theme
 export const DonutThresholdStaticTheme = {
   pie: {
-    colorScale: ['#EDEDED', '#D2D2D2', '#BBBBBB'],
-    height: 230,
-    padAngle: 1,
-    width: 230
+    colorScale: [
+      chart_donut_threshold_first_Color.value,
+      chart_donut_threshold_second_Color.value,
+      chart_donut_threshold_third_Color.value
+    ],
+    height: chart_donut_threshold_static_pie_Height.value,
+    padAngle: chart_donut_threshold_static_pie_angle_Padding.value,
+    width: chart_donut_threshold_static_pie_Width.value
   }
 };