+++ packages/patternfly-4/react-tokens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-tokens@2.6.5...@patternfly/react-tokens@2.6.6) (2019-07-03)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+
+
+
+
 ## [2.6.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-tokens@2.6.4...@patternfly/react-tokens@2.6.5) (2019-06-28)
 
 **Note:** Version bump only for package @patternfly/react-tokens